requirements.txt

typed-argument-parser>=1.8.1
python-dotenv>=1.0.0

# stuck with this version since I run python 3.12
# and this issue is unresolved: https://github.com/aio-libs/aiohttp/issues/7229
# furthermore, `py-cord` depends on `aiohttp<3.9.0`
aiohttp==3.9.0b0
discord.py>=2.3.2

# formatting
black>=22.3.0
